Vastly improve namespaced elements using subclasses, add database support for element attributes, feed_has_nselement and media_has_nselement have database changes.

+++ UNL/MediaYak/Models/BaseMediaHasNSElement.php
@@ -6,9 +6,16 @@ abstract class UNL_MediaYak_Models_BaseMediaHasNSElement extends Doctrine_Record
     public function setTableDefinition()
     {
         $this->setTableName('media_has_nselement');
-        $this->hasColumn('media_id',  'integer',    4, array('unsigned' => 0, 'primary' => true, 'notnull' => true, 'autoincrement' => true));
-        $this->hasColumn('nselement', 'string',  null, array('primary' => true, 'notnull' => true, 'autoincrement' => false));
-        $this->hasColumn('value',     'string',  null, array('primary' => false, 'notnull' => true, 'autoincrement' => false));
+        $this->hasColumn('media_id',   'integer',    4, array('unsigned' => 0, 'primary' => true, 'notnull' => true, 'autoincrement' => true));
+        $this->hasColumn('xmlns',      'string',  null, array('primary' => true, 'notnull' => true, 'autoincrement' => false));
+        $this->hasColumn('element',    'string',  null, array('primary' => true, 'notnull' => true, 'autoincrement' => false));
+        $this->hasColumn('attributes', 'string',  null, array('primary' => false, 'notnull' => false, 'autoincrement' => false));
+        $this->hasColumn('value',      'string',  null, array('primary' => false, 'notnull' => false, 'autoincrement' => false));
+        
+        $this->setSubclasses(array(
+                'UNL_MediaYak_Feed_Media_NamespacedElements_itunes' => array('xmlns' => 'itunes'),
+                'UNL_MediaYak_Feed_Media_NamespacedElements_mrss'   => array('xmlns' => 'mrss')
+            ));
     }
     
     public function setUp()
@@ -20,11 +27,7 @@ abstract class UNL_MediaYak_Models_BaseMediaHasNSElement extends Doctrine_Record
   
     function preInsert($event)
     {
-        if (!preg_match('/^'.$this->getXMLNS().':(.*)/', $this->nselement)) {
-            $event->skipOperation();
-            return;
-        }
-        if (empty($this->value)) {
+        if (empty($this->value) && empty($this->attributes)) {
             $event->skipOperation();
             return;
         }
@@ -32,11 +35,7 @@ abstract class UNL_MediaYak_Models_BaseMediaHasNSElement extends Doctrine_Record
     
     function preUpdate($event)
     {
-        if (!preg_match('/^'.$this->getXMLNS().':(.*)/', $this->nselement)) {
-            $event->skipOperation();
-            return;
-        }
-        if (empty($this->value)) {
+        if (empty($this->value) && empty($this->attributes)) {
             $this->delete();
             $event->skipOperation();
             return;
@@ -45,10 +44,7 @@ abstract class UNL_MediaYak_Models_BaseMediaHasNSElement extends Doctrine_Record
     
     function preDelete($event)
     {
-        if (!preg_match('/^'.$this->getXMLNS().':(.*)/', $this->nselement)) {
-            $event->skipOperation();
-            return;
-        }
+
     }
     
     /**
